Crude workarounds to make the unrealcv 4.22 branch compatible with UE 4.23
Some deprecated macros in TextureReader.cpp were replaced.
However IRendererModule::RenderTargetPoolFindFreeElement is problematic - I can't figure out how to replace it. Luckily, the ResizeFastReadTexture2dAsync doesn't seem to be called anywhere, so it has been entirely commented out with no observable ill affects.

In UnrealcvServer.cpp, the #if WITH_EDITOR statement fails when running in Standalone mode - that is, it evaluates to true, but the contents of it fail. The #else part of that statement has been added for if IsValid(EditorEngine) fails. Additional logs also added for debugging purposes - could be removed.